requirements.txt
numpy
matplotlib
streamlit
streamlit-drawable-canvas
pillow
